add new parser

+++ bot.py
@@ -1,4 +1,3 @@
-import asyncio
 import os
 import urllib
 
@@ -7,10 +6,15 @@ from aiogram import Dispatcher
 from aiogram.utils import executor
 from aiogram.contrib.fsm_storage.memory import MemoryStorage
 from deploy.config import TOKEN
-from parser import parser
+from new_parser import all_parsing
 
 bot = Bot(token=TOKEN)
 dp = Dispatcher(bot, storage=MemoryStorage())
+
+
+@dp.message_handler(commands='start')
+async def scan_message(msg: types.Message):
+    await bot.send_message(chat_id=msg.chat.id, text="Жду файл для обработки, формат файла xlsx")
 
 
 @dp.message_handler(content_types=['document'])
@@ -19,21 +23,36 @@ async def scan_message(msg: types.Message):
     file_info = await bot.get_file(document_id)
     fi = file_info.file_path
     file_name = msg.document.file_name
+    type_file = file_name.split('.')[-1]
     user_path = f'file/{msg.from_user.id}'
-    copy_file_name = 'copy' + file_name
-    os.makedirs(user_path, exist_ok=True)
-    original_file_path = urllib.request.urlretrieve(f'https://api.telegram.org/file/bot{TOKEN}/{fi}', f'{user_path}/{file_name}')
-    copy_file_path = urllib.request.urlretrieve(f'https://api.telegram.org/file/bot{TOKEN}/{fi}', f'{user_path}/{copy_file_name}')
-    await bot.send_message(msg.from_user.id, 'Начинаю парсинг')
-    copy_file = await parser(original_file_path[0], copy_file_path[0])
-    await bot.send_document(msg.from_user.id, open(copy_file, 'rb'))
-    os.remove(original_file_path[0])
-    os.remove(copy_file_path[0])
+    if type_file != 'xlsx':
+        await bot.send_message(msg.from_user.id, "Данный формат файла не поддердивается, работаем только с xlsx")
+    elif os.path.exists(f'{user_path}/{file_name}'):
+        await bot.send_message(msg.from_user.id, "Я уже работаю с этим файлом")
+    else:
+
+        try:
+            user_path = f'file/{msg.from_user.id}'
+            copy_file_name = 'copy' + file_name
+            os.makedirs(user_path, exist_ok=True)
+            original_file_path = urllib.request.urlretrieve(f'https://api.telegram.org/file/bot{TOKEN}/{fi}',
+                                                            f'{user_path}/{file_name}')
+            copy_file_path = urllib.request.urlretrieve(f'https://api.telegram.org/file/bot{TOKEN}/{fi}',
+                                                        f'{user_path}/{copy_file_name}')
+            await bot.send_message(msg.from_user.id, 'Начинаю парсинг')
+            copy_file = await all_parsing(original_file_path[0], copy_file_path[0])
+            await bot.send_document(msg.from_user.id, open(copy_file, 'rb'))
+        except Exception as e:
+            print(e)
+        finally:
+            os.remove(original_file_path[0])
+            os.remove(copy_file_path[0])
 
 
 @dp.message_handler()
 async def file_handle(message: types.message):
-    await bot.send_message(message.chat.id, message.text)
+    text = 'Это функция, для демонстрации асинхронной работы\nВаше сообщение\n' + message.text
+    await bot.send_message(message.chat.id, text)
 
 
 if __name__ == '__main__':
